backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.17.0
pydantic>=2.11.7,<3.0.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
pyjwt>=2.10.1,<3.0.0
httpx==0.27.0

